refactor(rust): cleanup `node create` arguments
